Adds configurability to more rules
Modifies the migrations check to be configurable,
also revises more of how the jira-test is configures.

Also cleans out old and unused parts of the config file
that have been replaced by the new jira configuration.

And adds notes in the roles.md for how to configure these
things.

Test Plan:
- tests pass